+++ db/api/views.py
@@ -1,16 +1,21 @@
 from rest_framework import viewsets
 
 from db.api import serializers, filters
-from db.base.models import Satellite, Transmitter
+from db.base.models import Mode, Satellite, Transmitter
 
 
-class SatelliteView(viewsets.ModelViewSet):
+class ModeView(viewsets.ReadOnlyModelViewSet):
+    queryset = Mode.objects.all()
+    serializer_class = serializers.ModeSerializer
+
+
+class SatelliteView(viewsets.ReadOnlyModelViewSet):
     queryset = Satellite.objects.all()
     serializer_class = serializers.SatelliteSerializer
     lookup_field = 'norad_cat_id'
 
 
-class TransmitterView(viewsets.ModelViewSet):
+class TransmitterView(viewsets.ReadOnlyModelViewSet):
     queryset = Transmitter.objects.all()
     serializer_class = serializers.TransmitterSerializer
     filter_class = filters.TransmitterViewFilter