+++ db/settings/base.py
@@ -96,6 +96,7 @@ STATICFILES_FINDERS = (
 MEDIA_ROOT = path.join(BASE_DIR, 'media')
 MEDIA_URL = '/media/'
 CRISPY_TEMPLATE_PACK = 'bootstrap3'
+SATELLITE_DEFAULT_IMAGE = '/static/img/sat.png'
 
 # App conf
 ROOT_URLCONF = 'db.urls'