+++ db/urls.py
@@ -28,4 +28,6 @@ if settings.DEBUG:
         '',
         url(r'^404/$', handler404),
         url(r'^500/$', handler500),
+        url(r'^media/(?P<path>.*)$', 'django.views.static.serve',
+            {'document_root': settings.MEDIA_ROOT}),
     )